Agentic_AI_Login3/pyproject.toml
[project]
name = "agentic-ai-login3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==3.2.2",
    "fastapi>=0.128.5",
    "passlib[bcrypt]==1.7.4",
    "pydantic[email]>=2.12.5",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
